requirements.txt
django
djangorestframework
django-simpleui
django-crontab
